Basic water drink reminder
Sometimes we are so busy in our work and we forget to perform or do the important work like eating or school assignments or even sleep if too busy. With all of these we sometimes also forget to drink water for that reason our body cannot get hydrated which form health issues so for that reason here's the basic code which remind us to drink water after input time by user. You just have to run the code simply and it work like best......